feat: optimization q001_two_sum

+++ src/q001_two_sum.rs
@@ -52,12 +52,12 @@ pub struct Solution {}
 //leetcode submit region begin(Prohibit modification and deletion)
 impl Solution {
     pub fn two_sum(nums: Vec<i32>, target: i32) -> Vec<i32> {
-        let mut map: HashMap<i32, i32> = HashMap::with_capacity(nums.len());
+        let mut map: HashMap<i32, usize> = HashMap::with_capacity(nums.len());
         for (i, num) in nums.iter().enumerate() {
             if let Some(&j) = map.get(&(target - num)) {
-                return vec![j, i as i32]
+                return vec![j as i32, i as i32]
             }
-            map.insert(*num, i as i32);
+            map.insert(*num, i);
         }
         vec![]
     }
